# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation: User To Do",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "User login",
                "description": "Logs in a user and returns user data with a token",
                "operationId": "loginUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "bhavik.test@yopmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "User12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User logged in successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 930952413
                                                },
                                                "first_name": {
                                                    "type": "string",
                                                    "example": "Bhavik"
                                                },
                                                "last_name": {
                                                    "type": "string",
                                                    "example": "Bhatiya"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "email",
                                                    "example": "bhavik.test@yopmail.com"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2024-06-29T04:48:31.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": "2024-06-29T04:48:31.000000Z"
                                                },
                                                "deleted_at": {
                                                    "type": "string",
                                                    "format": "datetime",
                                                    "example": null,
                                                    "nullable": true
                                                },
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGci....."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Email or password wrong."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "x_api_key": [],
                        "app_lang": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "x_api_key": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "X-API-Key",
                "in": "header"
            },
            "app_lang": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "app-language",
                "in": "header"
            },
            "passport": {
                "type": "http",
                "in": "header",
                "name": "bearer",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": {
        "1": {
            "x_api_key": [],
            "app_lang": [],
            "passport": []
        }
    }
}